+++ CHANGELOG.md
@@ -4,11 +4,25 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+# [1.39.0] - 2023-10-27
+
+### Added
+
+- Laiier - SevernWLD
+
+### Changed
+
+- Comtac - Cluey
+
 # [1.38.0] - 2023-09-28
+
+### Added
 
 - Nanothings - Nanotag
 
 # [1.38.0] - 2023-09-28
+
+### Added
 
 - Integra - Aquastream
 - Enginko - EGK-LW22PLG
@@ -16,8 +30,6 @@ All notable changes to this project will be documented in this file.
 - Enginko - R718N3
 - Dragino - Gropoint Air
 - Dragino - Gropoint Air V1.1
-
-### Added
 
 # [1.37.2] - 2023-08-30